@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace VasilDakov\Speedy\Exception;
 
@@ -11,5 +13,4 @@ namespace VasilDakov\Speedy\Exception;
  */
 class InvalidArgumentException extends \InvalidArgumentException
 {
-
 }
